refactor: dry up panels. refactor: modularize components of panels
